Moved colocated test into test/ because polyfill didnt work there
